+++ apps/server/src/routes/v1/(patch)/stations/[station_id]/cells/index.ts
@@ -235,6 +235,7 @@ async function handler(req: FastifyRequest<RequestData>, res: ReplyPayload<JSONB
     await Promise.all(
       response.map((cell) => {
         const oldFull = existingCells.find((c) => c.id === cell.id);
+        // oxlint-disable-next-line no-unused-vars: We want to remove those from the object
         const { gsm, umts, lte, nr, ...oldBase } = oldFull ?? ({} as typeof oldFull & Record<string, never>);
         const oldDetails = oldFull ? (oldFull.gsm ?? oldFull.umts ?? oldFull.lte ?? oldFull.nr ?? undefined) : undefined;
         return createAuditLog(